refactor: move BackgroundCluster to utils/distributed
Call-graph analysis shows BackgroundCluster has zero edges to any physics --
it is pure cluster-lifecycle infrastructure, yet compute.py imported it from
the cphotang physics module. Home it with the other cross-cutting utilities;
cphotang re-exports it for backward compatibility. No behavior change.
